Update to 5.6.0b11

+++ UnityEditor/ObjectInfo.cs
@@ -7,7 +7,7 @@
     {
         public string className;
         public int instanceId;
-        public int memorySize;
+        public long memorySize;
         public string name;
         public int reason;
         public List<ObjectInfo> referencedBy;
